+++ server/actions/index.js
@@ -1,0 +1,9 @@
+const notify = require("./notify-handler");
+const compute = require("./compute-handler");
+const statusCheck = require("./status-check");
+
+module.exports = {
+  notify,
+  compute,
+  statusCheck
+};