+++ server/is-group-trait.js
@@ -1,5 +1,0 @@
-import _ from "lodash";
-
-module.exports = function isGroup(o) {
-  return _.isPlainObject(o) && !_.isEqual(_.sortBy(_.keys(o)), ["operation", "value"]);
-};