+++ server/lib/utils/is-group-trait.js
@@ -1,0 +1,7 @@
+const _ = require("lodash");
+
+function isGroupTrait(hullObj) {
+  return _.isPlainObject(hullObj) && !_.isEqual(_.sortBy(_.keys(hullObj)), ["operation", "value"]);
+}
+
+module.exports = isGroupTrait;