+++ tests/integration/support/bootstrap.js
@@ -1,9 +1,11 @@
 const Hull = require("hull");
-const server = require("../../../server/server").default;
+const server = require("../../../server/server");
 
 module.exports = function bootstrap() {
   const hostSecret = "1234";
-  const connector = new Hull.Connector({ hostSecret, port: 8000, skipSignatureValidation: true, clientConfig: { protocol: "http", firehoseUrl: "firehose" } });
+  const connector = new Hull.Connector({
+    hostSecret, port: 8000, skipSignatureValidation: true, clientConfig: { protocol: "http", firehoseUrl: "firehose" }
+  });
   const app = server(connector, {
     Hull,
     hostSecret