+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## 0.2.14
+
+- [improvement] Upgrade code to node v8.9.x LTS release
+- [improvement] Use customizable tabs for code editor
+- [bugfix] Fix failing code syntax check
+- [bugfix] Fix failing initial user load
+- [maintenance] Re-factoring of code base for better readibility and maintainability
+- [maintenance] Additional tests for linting and syntax checking of more complex code
+- [maintenance] Exclude selected events from `events` array (see README.md Notes section for details)
+
 ## 0.2.13
 
 - Don't display Attributes changed, Segments changed, Entered & Left segments events in the preview UI